We now have client-side feedback upon changing and deleting
