+++ views/layouts/main.handlebars
@@ -6,6 +6,7 @@
     <link rel="stylesheet" href="/main.css">
     <script src="/style.js" defer></script>
     <script src="/ux.js" defer></script>
+    <script src="/ENTITIES.js"></script>
     <script>
         let addAbilitiesForm = function ()
         {
